// src/types/TProperty.ts
export type TProperty = {
  id: number;
  title: string;
  description: string;
  location: string;
  source: string;
  url: string;
  price_value: string;
  price_currency: string;
  price_period: string;
  property_type: string;
  property_purpose: string;
  reference_no: string;
  completion_status: string | null;
  added_date: Date;
  bedrooms: number | null;
  bathrooms: number | null;
  size_value: string;
  size_unit: string;
  agent_name: string | null;
  agent_image_url: string | null;
  amenities: string[];
  image_urls: string[];
  verified: boolean | null;
  agent_number: string;
  created_at: Date;
  whatsapp_number: string;
  latitude: string;
  longitude: string;
  is_approved: boolean;
};